Make Client.drain() throw InterruptedException. It no longer spins. Simply VoltNetwork.shutdown(). Fix ENG-558 by modifying VoltProcedure. Fix ENG-537. VoltCompiler now requires procedures partition parameters defined for procedures to have the same type as the partition column.
Fix broken examples.

+++ examples/auction/procedures/com/auctionexample/AuctionStatus.java
@@ -73,7 +73,7 @@ public class AuctionStatus extends VoltProcedure {
      *  (item, bidder, seller, bidcount, price, endtime, status)
      * @throws VoltAbortException Currently doesn't abort.
      */
-    public VoltTable[] run(long itemId) throws VoltAbortException {
+    public VoltTable[] run(int itemId) throws VoltAbortException {
         // create a new VoltTable to store our results
         VoltTable retval = new VoltTable(
                 new VoltTable.ColumnInfo("item", VoltType.STRING),
