+++ src/proj_gen/Delete.java
@@ -27,7 +27,7 @@ public class Delete extends VoltProcedure {
       new SQLStmt("DELETE FROM ##upper_project_name## " +
                   "WHERE ##upper_project_name##_ID = ?");
 
-    public long run( long ##upper_project_name##_ID ) throws VoltAbortException {
+    public long run( int ##upper_project_name##_ID ) throws VoltAbortException {
         // Add a SQL statement to the execution queue.
         voltQueueSQL( deleteItem, ##upper_project_name##_ID );
