+++ tests/frontend/org/voltdb/benchmark/tpcc/procedures/InsertNewOrder.java
@@ -60,7 +60,7 @@ public class InsertNewOrder extends VoltProcedure {
 
     public final SQLStmt insert = new SQLStmt("INSERT INTO NEW_ORDER VALUES (?, ?, ?);");
 
-    public VoltTable[] run(long no_o_id, long no_d_id, long no_w_id) {
+    public VoltTable[] run(long no_o_id, long no_d_id, short no_w_id) {
         voltQueueSQL(insert, no_o_id, no_d_id, no_w_id);
         return voltExecuteSQL();
     }
